42323ing

+++ code/IOICAMP/2025/Day4/A.cpp
@@ -65,7 +65,7 @@ signed main() {
     int l=1,r=2e9,m;
     while(l<r) {
         m=l+r+1>>1;
-        __int128 cnt=0;
+        int cnt=0;
         REP(i,n) if(a[i]<m) cnt+=(m-a[i]+c-1)/c;
         op(l)op(r)op(m)ope(cnt)
         if(cnt<=m) r=m-1;
